# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 6263e770c97c70218d65385c
  title: SMS
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/class/{id}":
    delete:
      x-skaffolder-id: 6263f568c97c70218d653939
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 6263f513c97c70218d65391e
      x-skaffolder-resource: Class
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles:
        - 6263e771c97c70218d653868
      summary: CRUD ACTION delete
      tags:
        - Class
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 6263f568c97c70218d653932
      x-skaffolder-name: get
      x-skaffolder-id-resource: 6263f513c97c70218d65391e
      x-skaffolder-resource: Class
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Class
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    post:
      x-skaffolder-id: 6263f568c97c70218d653934
      x-skaffolder-name: update
      x-skaffolder-id-resource: 6263f513c97c70218d65391e
      x-skaffolder-resource: Class
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles:
        - 6263e771c97c70218d653868
      summary: CRUD ACTION update
      tags:
        - Class
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Class"
  "/class/findBy_Student/{key}":
    get:
      x-skaffolder-id: 6263f538c97c70218d653924
      x-skaffolder-name: findBy_Student
      x-skaffolder-id-resource: 6263f513c97c70218d65391e
      x-skaffolder-resource: Class
      x-skaffolder-crudAction: findBy_Student
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_Student
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_Student/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_Student
      tags:
        - Class
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
  "/class/findBy_Teacher/{key}":
    get:
      x-skaffolder-id: 6263f547c97c70218d65392d
      x-skaffolder-name: findBy_Teacher
      x-skaffolder-id-resource: 6263f513c97c70218d65391e
      x-skaffolder-resource: Class
      x-skaffolder-crudAction: findBy_Teacher
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_Teacher
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_Teacher/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_Teacher
      tags:
        - Class
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
...
  "/student/{id}":
    delete:
      x-skaffolder-id: 6263eb77c97c70218d6538f8
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 6263ea05c97c70218d6538f3
      x-skaffolder-resource: Student
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles:
        - 6263e771c97c70218d653868
      summary: CRUD ACTION delete
      tags:
        - Student
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 6263f538c97c70218d653921
      x-skaffolder-name: get
      x-skaffolder-id-resource: 6263ea05c97c70218d6538f3
      x-skaffolder-resource: Student
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    post:
      x-skaffolder-id: 6263eb8cc97c70218d6538f9
      x-skaffolder-name: update
      x-skaffolder-id-resource: 6263ea05c97c70218d6538f3
      x-skaffolder-resource: Student
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: Student
      x-skaffolder-url: /{id}
      x-skaffolder-roles:
        - 6263e771c97c70218d653868
      summary: CRUD ACTION update
      tags:
        - Student
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Student"
...
  "/student/findBy_Teacher/{key}":
    get:
      x-skaffolder-id: 6263f547c97c70218d65392b
      x-skaffolder-name: findBy_Teacher
      x-skaffolder-id-resource: 6263ea05c97c70218d6538f3
      x-skaffolder-resource: Student
      x-skaffolder-crudAction: findBy_Teacher
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_Teacher
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_Teacher/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_Teacher
      tags:
        - Student
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Class:
      x-skaffolder-id: 6263f513c97c70218d65391e
      x-skaffolder-id-db: 6263e770c97c70218d65385d
      x-skaffolder-id-entity: 6263f512c97c70218d653919
      x-skaffolder-url: /class
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        ClassName:
          type: number
          x-skaffolder-id-attr: 6263f512c97c70218d65391a
          x-skaffolder-type: Number
          x-skaffolder-required: true
        Section:
          type: string
          x-skaffolder-id-attr: 6263f512c97c70218d65391b
          x-skaffolder-type: String
      x-skaffolder-relations:
        _Student:
          x-skaffolder-id: 6263f512c97c70218d65391c
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 6263f512c97c70218d653919
          x-skaffolder-ent2: 6263ea05c97c70218d6538ec
        _Teacher:
          x-skaffolder-id: 6263f512c97c70218d65391d
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 6263f512c97c70218d653919
          x-skaffolder-ent2: 6263ed3cc97c70218d653900
      required:
        - _id
        - ClassName
    Student:
      x-skaffolder-id: 6263ea05c97c70218d6538f3
      x-skaffolder-id-db: 6263e770c97c70218d65385d
      x-skaffolder-id-entity: 6263ea05c97c70218d6538ec
      x-skaffolder-url: /student
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Class:
          type: number
          x-skaffolder-id-attr: 6263ea05c97c70218d6538f2
          x-skaffolder-type: Number
          x-skaffolder-required: true
        FathersName:
          type: string
          x-skaffolder-id-attr: 6263ea05c97c70218d6538f0
          x-skaffolder-type: String
        FirstName:
          type: string
          x-skaffolder-id-attr: 6263ea05c97c70218d6538ed
          x-skaffolder-type: String
          x-skaffolder-required: true
        LastName:
          type: string
          x-skaffolder-id-attr: 6263ea05c97c70218d6538ef
          x-skaffolder-type: String
          x-skaffolder-required: true
        MiddleName:
          type: string
          x-skaffolder-id-attr: 6263ea05c97c70218d6538ee
          x-skaffolder-type: String
        MothersName:
          type: string
          x-skaffolder-id-attr: 6263ea05c97c70218d6538f1
          x-skaffolder-type: String
        RollNumber:
          type: object
          x-skaffolder-id-attr: 6263ec25c97c70218d6538fd
          x-skaffolder-type: Custom
          x-skaffolder-required: true
          x-skaffolder-unique: true
      x-skaffolder-relations:
        _Teacher:
          x-skaffolder-id: 6263ee76c97c70218d65390d
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 6263ea05c97c70218d6538ec
          x-skaffolder-ent2: 6263ed3cc97c70218d653900
      required:
        - _id
        - Class
        - FirstName
        - LastName
        - RollNumber
...